+++ neo/NeoPlugin.php
@@ -22,7 +22,7 @@ class NeoPlugin extends BasePlugin
 
 	public function getVersion()
 	{
-		return '1.5.1';
+		return '1.5.2';
 	}
 
 	public function getCraftMinimumVersion()